requirements.txt
aiofiles==0.5.0
click==7.1.2
fastapi==0.59.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.6.1
PyJWT==1.7.1
python-dotenv==0.14.0
python-multipart==0.0.5
six==1.15.0
starlette==0.13.4
uvicorn==0.11.6
uvloop==0.14.0
websockets==8.1
